requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
async-timeout==4.0.2
attrs==24.2.0
audioop-lts==0.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
yarl==1.18.3
